assembly renaming and document formatting
