+++ Lmbtfy.UnitTests/Lmbtfy.UnitTests.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Lmbtfy.Web\Lmbtfy.Web.csproj" />
+  </ItemGroup>
+
 </Project>
