+++ Lmbtfy.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 15.0.28307.168
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lmbtfy.Web", "Lmbtfy.Web\Lmbtfy.Web.csproj", "{EB5F205E-1C55-4E2F-99F8-1E97BF6F285F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests", "UnitTests\UnitTests.csproj", "{79FC62C3-0655-4E59-B142-B505454D42ED}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lmbtfy.UnitTests", "Lmbtfy.UnitTests\Lmbtfy.UnitTests.csproj", "{105BCCA6-CA32-4BA7-BBC4-5E300706AFB4}"
 EndProject
 Global
@@ -19,10 +17,6 @@ Global
 		{EB5F205E-1C55-4E2F-99F8-1E97BF6F285F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EB5F205E-1C55-4E2F-99F8-1E97BF6F285F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EB5F205E-1C55-4E2F-99F8-1E97BF6F285F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{79FC62C3-0655-4E59-B142-B505454D42ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{79FC62C3-0655-4E59-B142-B505454D42ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{79FC62C3-0655-4E59-B142-B505454D42ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{79FC62C3-0655-4E59-B142-B505454D42ED}.Release|Any CPU.Build.0 = Release|Any CPU
 		{105BCCA6-CA32-4BA7-BBC4-5E300706AFB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{105BCCA6-CA32-4BA7-BBC4-5E300706AFB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{105BCCA6-CA32-4BA7-BBC4-5E300706AFB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
